% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_class_02_klausuR.R
\docType{class}
\name{klausuR,-class}
\alias{klausuR,-class}
\alias{klausuR-class}
\title{Class klausuR}
\description{
This class is used for objects that are returned by \code{\link[klausuR:klausur]{klausur}}.
}
\section{Slots}{

\describe{
\item{\code{results}}{A data.frame with global results}

\item{\code{answ}}{A data.frame with all given answers}

\item{\code{corr}}{A vector with the correct answers}

\item{\code{wght}}{A vector with the weights of items}

\item{\code{points}}{A data.frame with resulting points given for the answers}

\item{\code{marks}}{A vector with assignments of marks to achieved score}

\item{\code{marks.sum}}{A more convenient matrix with summary information on the defined marks}

\item{\code{trfls}}{A data.frame of TRUE/FALSE values,
      whether a subject was able to solve an item or not}

\item{\code{anon}}{A data.frame for anonymous feedback}

\item{\code{mean}}{A table with mean, median and quartiles of the test results}

\item{\code{sd}}{Standard deviation of the test results}

\item{\code{cronbach}}{Internal consistency, a list of three elements "alpha",
      "ci" (confidence interval 95\%) and "deleted" (alpha if item was removed)}

\item{\code{item.analysis}}{A data.frame with information on difficulty,
      discriminant power, discriminant factor and selection index of all items.}

\item{\code{distractor.analysis}}{A list with information on the selected answer alternatives for each individual item.}

\item{\code{test}}{Currently an empty placeholder. Planned to hold the actual test items in future releases.}

\item{\code{misc}}{Anything that was stored in the \code{misc} slot of the input data.}
}}
\keyword{classes}

