% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klausuR-package.R
\docType{package}
\name{klausuR-package}
\alias{klausuR}
\alias{klausuR-package}
\title{\packageTitle{klausuR}}
\description{
\packageDescription{klausuR}
}
\details{
The DESCRIPTION file:
\tabular{ll}{
Package: \tab klausuR\cr
Type: \tab Package\cr
Version: \tab 0.12-15\cr
Date: \tab 2023-06-12\cr
Depends: \tab R (>= 2.9.0), xtable\cr
Enhances: \tab rkward\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab https://reaktanz.de/?c=hacking&s=klausuR\cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://reaktanz.de/?c=hacking&s=klausuR}
  \item Report bugs at \url{https://github.com/unDocUMeantIt/klausuR/issues}
}

}
\author{
\packageAuthor{klausuR}

Maintainer: \packageMaintainer{klausuR}
}
