\name{klausuR.mult-class}
\alias{klausuR.mult-class}
\alias{klausuR.mult,-class
-class}
\alias{klausuR.mult,-class}
\title{S4 class klausuR.mult}
\description{Class klausuR.mult}
\details{This class is used for objects that are returned by \code{\link[klausuR:klausur.mufo]{klausur.mufo}}.}
\keyword{classes}
\author{m.eik michalke \email{meik.michalke@uni-duesseldorf.de}}
\section{Slots}{\describe{\item{\code{forms}:}{(\code{\link{vector}}) A vector with the names of all test forms.}\item{\code{results.part}:}{(\code{\link{list}}) A list with the partial results of each test form}\item{\code{results.glob}:}{(\code{\link{klausuR}}) An object of class klausuR-class with overall results}}}

