\name{show}
\alias{show,-method}
\alias{show}
\title{show method for objects of class klausuR}
\description{show method for S4 objects of class klausuR}
\details{Prints a nice summary of the slots \code{results}, \code{anon}, \code{cronbach} and \code{item.analysis}.}
\alias{show,klausuR-method}
\author{m.eik michalke \email{meik.michalke@uni-duesseldorf.de}}
\seealso{\code{\link[klausuR:klausur]{klausur}}}
\keyword{methods}
\arguments{\item{object}{An object of class klausuR}
}
\examples{\dontrun{
klausur(answ=antworten, corr=richtig, marks=notenschluessel)
}}

