% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/klausuR-package.R
\docType{package}
\name{klausuR-package}
\alias{klausuR-package}
\title{The klausuR Package}
\description{
Multiple Choice Test Evaluation.
}
\details{
\tabular{ll}{
Package: \tab klausuR\cr
Type: \tab Package\cr
Version: \tab 0.12-10\cr
Date: \tab 2015-02-07\cr
Depends: \tab R (>= 2.9.0), xtable, psychometric, polycor, methods, graphics, tools\cr
Enhances: \tab rkward\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab http://r-forge.r-project.org/projects/klausur\cr
}

A set of functions designed to quickly generate results of
a multiple choice test. Generates detailed global results, lists
for anonymous feedback and personalised result feedback (in LaTeX
and/or PDF format), as well as item statistics like Cronbach's alpha or
disciminatory power. klausuR also includes a plugin for the R GUI and
IDE RKWard, providing dialogs for its basic features. To use them,
install RKWard from http://rkward.sf.net (plugins are detected
automatically). Due to some restrictions on CRAN, the full package
sources are only available from the project homepage.
}
\author{
m.eik michalke
}
\keyword{package}

