\docType{package}
\name{klausuR-package}
\alias{klausuR-package}
\title{The klausuR Package}
\description{
  Multiple Choice Test Evaluation.
}
\details{
  \tabular{ll}{ Package: \tab klausuR\cr Type: \tab
  Package\cr Version: \tab 0.12-9\cr Date: \tab
  2013-08-05\cr Depends: \tab R (>=
  2.9.0),xtable,psychometric,methods,graphics,tools\cr
  Enhances: \tab rkward\cr Encoding: \tab UTF-8\cr License:
  \tab GPL (>= 3)\cr LazyLoad: \tab yes\cr URL: \tab
  http://r-forge.r-project.org/projects/klausur\cr }

  A set of functions designed to quickly generate results
  of a multiple choice test. Generates detailed global
  results, lists for anonymous feedback and personalised
  result feedback (in LaTeX and/or PDF format), as well as
  item statistics like Cronbach's alpha or disciminatory
  power.
}
\author{
  Meik Michalke \email{meik.michalke@hhu.de}
}
\keyword{package}

