% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/00_class_04_klausuR.mult.R
\docType{class}
\name{klausuR.mult-class}
\alias{klausuR.mult,-class}
\alias{klausuR.mult-class}
\title{Class klausuR.mult}
\description{
This class is used for objects that are returned by \code{\link[klausuR:klausur.mufo]{klausur.mufo}}.
}
\section{Slots}{

\describe{
\item{\code{forms}}{A vector with the names of all test forms.}

\item{\code{results.part}}{A list with the partial results of each test form}

\item{\code{results.glob}}{An object of class klausuR-class with overall results}
}}
\keyword{classes}

