\docType{class}
\name{klausuR.test,-class}
\alias{klausuR.test,-class}
\alias{klausuR.test-class}
\title{Class klausuR.test}
\description{
This class is currently an empty placeholder. In future
releases, it is planned to contain the actual test items in
a list like format.
}
\section{Slots}{

\describe{
\item{\code{items}}{Empty dummy.}
}}
\keyword{classes}

