\name{klausuR.test-class}
\alias{klausuR.test,-class
-class}
\alias{klausuR.test,-class}
\title{S4 class klausuR.test}
\description{Class klausuR.test}
\details{This class is currently an empty placeholder. In future releases, it is planned to contain the actual test items in a list
like format.}
\keyword{classes}
\author{m.eik michalke \email{meik.michalke@uni-duesseldorf.de}}
\section{Slots}{\describe{\item{\code{items}:}{(\code{\link{list}}) Empty dummy.}}}

