% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klausur_meta.R
\name{klausur_meta}
\alias{klausur_meta}
\title{Arrange meta data for a test to analyze}
\usage{
klausur_meta(
  title,
  name,
  data_dir,
  data_file,
  corr,
  marks,
  ignore,
  subdir_prefix,
  rename = c(),
  dummies = c(),
  disc.misc = FALSE,
  na.rm = TRUE,
  item.prefix = c(),
  out_dir = data_dir,
  date = Sys.Date(),
  date_print = "\%d.\%m.\%Y",
  gt_suffix = "all.xlsx",
  create_out_dir = FALSE,
  ...
)
}
\arguments{
\item{title}{Character string, title of the test.}

\item{name}{Character string, name of the lecturer applying the test.}

\item{data_dir}{Character string, root directory for test data.}

\item{data_file}{File name of the test raw data. Use a path releative to \code{data_dir}. If this argument
is defined,
       \code{\link[klausuR:klausur.data]{klausur.data}} will try to import this file using
\code{\link[rio:import]{rio::import}}.}

\item{corr}{Named vector,
      see \code{corr} argument of \code{\link[klausuR:klausur.data]{klausur.data}}.}

\item{marks}{Character vector,
      see \code{marks} argument of \code{\link[klausuR:klausur.data]{klausur.data}}, and
\code{\link[klausuR:klausur.gen.marks]{klausur.gen.marks}}.
If you provide a named numeric vector,
      the names will be used as mark labels and the numbers as the maximum number of
points for that note. Start with the worst mark (test failed).}

\item{ignore}{Optional character vector,
      can be used as the \code{ignore} vector of \code{\link[klausuR:compare]{compare}}.}

\item{subdir_prefix}{Optional character string. If set,
      a subdirectory named \code{<subdir_prefix>_<date>} will be
added to both \code{data_dir} and \code{out_dir},
      with \code{<date>} being in \code{YYY-MM-DD} format. If \code{out_dir}
is the same as \code{data_dir}, it will have this subdirectory added twice recursively,
      so reports are easier to archive.}

\item{rename}{Optional named character vector to use with \code{rename} (see \code{\link[klausuR:klausur.data]{klausur.data}}) if the \code{No},
      \code{Name}, \code{FirstName}, \code{MatrNo}, \code{Pseudonym},
      or \code{Form} column (see \code{\link[klausuR:klausur.mufo]{klausur.mufo}}) has a different name in your data set.}

\item{dummies}{Optional character vector,
      see \code{dummies} argument of \code{\link[klausuR:klausur.data]{klausur.data}}.}

\item{disc.misc}{Logical, see \code{\link[klausuR:klausur.data]{klausur.data}}.}

\item{na.rm}{Logical, see \code{\link[klausuR:klausur.data]{klausur.data}}.}

\item{item.prefix}{Character, see \code{\link[klausuR:klausur.data]{klausur.data}}.}

\item{out_dir}{Character string, root directory for test results.}

\item{date_print}{Character string,
      defines the date format used in generated PDF documents.}

\item{gt_suffix}{Character string,
      suggests a file name for saving the \code{\link[klausuR:grand.table]{grand.table}} in a format supported by \code{\link[rio:export]{rio::export}}.}

\item{create_out_dir}{Logical, if \code{TRUE} and \code{out_dir} is missing,
      it will automatically be created.}

\item{...}{Optional additional arguments will be kept as-is in the resulting list.}
}
\value{
A named list with entries \code{title}, \code{name}, \code{date}, \code{data_dir},
      \code{out_dir}, \code{subdir}, \code{corr}, 
   \code{date}, \code{date_ISO}, \code{date_print}, \code{rename}, \code{dummies},
      \code{marks}, and \code{gt_file}.
}
\description{
Creates a list with formatted entries that can be used by functions like \code{\link[klausuR:klausur.data]{klausur.data}}.
}
