% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_method_report.klausuR.R
\docType{methods}
\name{report}
\alias{report}
\alias{report,klausuR-method}
\title{Create reports from evaluated tests}
\usage{
report(
  klsr,
  matn,
  path,
  file_name = "MatrNo",
  also_valid = c("_", "-"),
  save = FALSE,
  pdf = FALSE,
  body,
  template,
  header,
  ...
)

\S4method{report}{klausuR}(
  klsr,
  matn,
  path,
  file_name = "MatrNo",
  also_valid = c("_", "-"),
  save = FALSE,
  pdf = FALSE,
  body,
  template,
  header,
  ...
)
}
\arguments{
\item{klsr}{An object of class \code{klausuR}.}

\item{matn}{TODO}

\item{path}{TODO}

\item{file_name}{Character vector,
      either defining the output file name directly (if \code{matn="anon"} or \code{matn="glob"}),
      or defining the column names of the results slot in \code{klsr} whose
values should be used to create individual file names.}

\item{also_valid}{Character vector,
      values for \code{file_name} that should also be considered valid even if not column names of the results slot in \code{klsr}. They will be used as-is.}

\item{save}{TODO}

\item{pdf}{TODO}

\item{body}{TODO}

\item{template}{TODO}

\item{header}{A list of variables to be defined in the YAML header of each RMarkdown file to be available for use in \code{template}.
Lists of lists are supported, see \code{\link[yaml:as_yaml]{yaml::as_yaml}} for details.
For individual reports,
      \code{header} will get a nested variable named \code{results} including all variables/columns of the results slot in \code{klsr} with their respective values
of the row matching \code{matn}.}

\item{...}{TODO}
}
\description{
This is a rewrite of \code{klausur.report()} using RMarkdown instead of \LaTeX{}.
}
\details{
If you want to use a certain value from the results slot in \code{klsr} in the code of \code{body},
      you can use \code{\\newcommand} in your \code{template} and use the command in the markdown.
}
