% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_method_report.klausuR.R
\docType{methods}
\name{report}
\alias{report}
\alias{report,klausuR-method}
\title{Create reports from evaluated tests}
\usage{
report(
  klsr,
  matn,
  path,
  file_name = "MatrNo",
  also_valid = c("_", "-"),
  save = FALSE,
  pdf = FALSE,
  body,
  template,
  header,
  statistics = c(hist_points = FALSE, hist_marks = FALSE, marks_points = FALSE,
    marks_percent = FALSE),
  plot_names = c(hist_points = "hist_points.pdf", hist_marks = "hist_marks.pdf"),
  meta,
  ...
)

\S4method{report}{klausuR}(
  klsr,
  matn,
  path,
  file_name = "MatrNo",
  also_valid = c("_", "-"),
  save = FALSE,
  pdf = FALSE,
  body,
  template,
  header,
  statistics = c(hist_points = FALSE, hist_marks = FALSE, marks_points = FALSE,
    marks_percent = FALSE),
  plot_names = c(hist_points = "hist_points.pdf", hist_marks = "hist_marks.pdf"),
  meta,
  ...
)
}
\arguments{
\item{klsr}{An object of class \code{klausuR}.}

\item{matn}{TODO}

\item{path}{TODO}

\item{file_name}{Character vector,
      either defining the output file name directly (if \code{matn="anon"} or \code{matn="glob"}),
      or defining the column names of the results slot in \code{klsr} whose
values should be used to create individual file names.}

\item{also_valid}{Character vector,
      values for \code{file_name} that should also be considered valid even if not column names of the results slot in \code{klsr}. They will be used as-is.}

\item{save}{TODO}

\item{pdf}{TODO}

\item{body}{TODO}

\item{template}{TODO}

\item{header}{A list of variables to be defined in the YAML header of each RMarkdown file to be available for use in \code{template}.
Lists of lists are supported, see \code{\link[yaml:as_yaml]{yaml::as_yaml}} for details.
For individual reports,
      \code{header} will get a nested variable named \code{results} including all variables/columns of the results slot in \code{klsr} with their respective values
of the row matching \code{matn}. It will also include the \code{meta} values named \code{lang} and \code{labels}.}

\item{statistics}{Logical vector with recognized elements \code{hist_points},
      \code{hist_marks}, \code{marks_points}, and \code{marks_percent},
      indicating whether history plots for the
distribution of points and marks or a table showing the marks grading scale should be included,
      including points, percent, or both. Missing elements default to \code{FALSE}.}

\item{plot_names}{Character vector with recognized entries \code{hist_points} and \code{hist_marks},
      defining the file names for the histogram of points and marks respectively.}

\item{meta}{A list generated by \code{\link[klausuR:klausur_meta]{klausur_meta}}. These elements are directly used in generated report documents if present:
\itemize{
  \item{\code{title}: title of the test}
  \item{\code{name}: name of the lecturer applying the test}
  \item{\code{date_print}: the formatted date string}
  \item{\code{lang}: language to set internal defaults for \code{labels} (see below); set to "de" for reports in German,
      English is the default}
  \item{\code{labels}: list with recognized elements \code{p_main}, \code{p_xlab},
      \code{p_ylab}, \code{m_main}, \code{m_xlab}, and \code{m_ylab},
    defining the main title and axis annotation of points and marks histograms,
      respectively; \code{points}, \code{percent}, \code{m_scale} and \code{dist}
    defining the labels to use for points, percent,
      marks grading scale and distribution of test results}
}}

\item{...}{TODO}
}
\description{
This is a rewrite of \code{klausur.report()} using RMarkdown instead of \LaTeX{}.
}
\details{
If you want to use a certain value from the results slot in \code{klsr} in the code of \code{body},
      you can use \code{\\newcommand} in your \code{template} and use the command in the markdown.
}
