% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_method_show.klausuR.R
\docType{methods}
\name{show,klausuR-method}
\alias{show,-methods}
\alias{show,klausuR-method}
\alias{show,klausuR.mult-method}
\title{Show methods for S4 objects of classes klausuR and klausuR.mult}
\usage{
\S4method{show}{klausuR}(object)

\S4method{show}{klausuR.mult}(object)
}
\arguments{
\item{object}{An object of class \code{klausuR} or \code{klausuR.mult}}
}
\description{
Print a nice summary of the slots \code{results}, \code{anon},
      \code{cronbach} and \code{item.analysis}.
If object is of class klausuR.mult,
      the global results for tests with several test forms are evaluated.
}
\examples{
\dontrun{
klausur(data.obj)

# multiple test forms
klausur.mufo(data.obj, marks=notenschluessel)
}
}
\author{
m.eik michalke \email{meik.michalke@uni-duesseldorf.de}
}
\seealso{
\code{\link[klausuR:klausur]{klausur}}, \code{\link[klausuR:klausur.mufo]{klausur.mufo}}
}
\keyword{methods}

